htr_project/recognition/ml_models/requirements.txt
django==5.0
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
matplotlib==3.8.2

